+++ oop/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "functional-domain-modeling",
+  "name": "oop-domain-modeling",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
